Frontend:   New pages has been added.   Logic has been improved.   Styles has been improved. Backend:   Dao for some requests has been added.   Enteties has been changed.   New logic for services and repositories has been added.   Profile controller has been added.

+++ src/main/java/reseacrh/project/students/repositories/BacklinkingRepo.java
@@ -19,5 +19,5 @@ public interface BacklinkingRepo extends JpaRepository<Backlinking, Integer> {
     List<Backlinking> findCandidates(Long ownerId);
 
     @Query(value = "SELECT publication_id FROM backlinking WHERE candidate=?1", nativeQuery = true)
-    List<String> findAppliedPublicationsByUserId(Integer id);
+    List<String> findAppliedPublicationsByUserId(Long id);
 }
